requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django==5.2.9
gunicorn==23.0.0
pillow==12.0.0
psycopg[binary]==3.2.9
python-dotenv==1.2.1
reportlab==4.4.7
sqlparse==0.5.5
